C++ conversion part 2
remove most of the remaining C-isms from plugin/modland/hacks sources and convert to proper c++
