+++ configure.ac
@@ -1,6 +1,5 @@
 AC_PREREQ(2.59)
 AC_INIT([audacious-uade], [0.5-dev])
-AC_PROG_CC
 AC_PROG_CXX
 AX_CXX_COMPILE_STDCXX_17
